Fix build failure
Build failed during test phase with this error:

```
Dependency helm failed to install: Package `emacs-24.4' is unavailable
```

The current version of [Helm in MELPA stable](https://stable.melpa.org/#/helm)
requires emacs-24.4.  The build script installs emacs24-nox found in
[ppa:cassou/emacs](https://launchpad.net/~cassou/+archive/ubuntu/emacs) which
appears to be deprecated already and has only up to 24.3.  [Searching for the
emacs24 package in any Ubuntu
version](https://packages.ubuntu.com/search?suite=default&section=all&arch=any&keywords=emacs24&searchon=names)
reveals that the closest version available currently is 24.5 in Ubuntu 16.04.
Only Ubuntu 12.04 and 14.04 are available to the current CircleCI 1.0 build
script.  It must be upgraded to 2.0 in order to be able to build in a 16.04 environment.